OPSMN-5527: Improved the specimens retrieve API to perform better irrespective of the number of visit specimens.
The problem was - the collection and received events of primary specimens
were retrieved one at a time. This meant a visit with 500 primary specimens
resulted in 1000 SQLs.

Fixed the specimen read APIs to fetch the collection and received
events in batches of 100. This means a visit with 500 primary specimens
will result in only 5 DB calls.

Dev testing observation on a worn out laptop showed an improvement of 60x.
Time to load the visit page of 500 primary specimens reduced from ~9 mins
to 9 secs.