OPSMN-5524: Update DE library with the following query fix.
---

OPSMN-5524: Break long IN and NOT IN clauses into multiple manageable,
acceptable fragments.

Oracle does not permit IN and NOT IN clauses to have more than 1000 literals.

Therefore long IN and NOT IN clauses are broken down as below:

Example 1:

input:
IN (a1, a2,... a1001, a1002,... a2000)

output:
IN (a1, a2,... a1000) or IN (a1001, a1002,... a2000)

Example 2:

input:
NOT IN (a1, a2,... a1001, a1002,... a2000)

output:
NOT (IN (a1, a2,... a1000) or IN (a1001, a1002,... a2000))