AngularJS upgrade regression fix.
The toggle checkboxes in the query define view were not responding
to the user choices. The view controller was listening for the
click events, which are fired before the model value changes.
As a result, a tick was resulting in untick and vice-versa.

Updated the controller to listen for the  model changes rather
than the click events.