OPSMN-5632: Specimen collection workflow improvements
1. Init the collection date to current date and time.

2. Init the collector to either the one specified in the
corresponding specimen requirement or current user.

3. Made collector a mandatory field. This means users can no
longer remove it by clicking on x icon button.

4. Hyperlinked the specimen description to the corresponding
overview page. The behaviour is same as that of clicking link
on any specimen in the visits' specimens tree.

5. Displayed the visit details at all times.

6. Fixed the Submit & Print button to download the manifest when all
the specimens of the visit are precollected.

7. Fixed the timepicker widget to display the time instead of random
strings. This happened when user switched between the events on the
collection page.