OPSMN-5655: Added a new specimen label token that generates child specimen labels by incrementing the number that appears at the end of the primary specimen (specimen drawn from the patient) label.
Token: %PRIMARY_SPEC_COUNTER%

Added an optional numeric argument, which specifies the number of
digits of the primary specimen that needs to be fixed/constant or
should be carried over to all the child specimens.

%PRIMARY_SPEC_COUNTER(1)%

Fixes the first digit of the number appearing at the end of the primary
specimen label.

For example: If WB.10 is the primary specimen label then the child
specimen labels for the format %PRIMARY_SPEC_COUNTER(1)% will be:

Se.11, Se.12, ... Se.19, Se.110, Se.111, ..., Se.120 etc

Updated %PSPEC_COUNTER% to support the fixed digits argument.