OPSMN-5662: Solved the circular dependencies in the application beans/services.
1. Password and other secure strings are encrypted and stored in the database.
(The database is encrypted but the strings are further encrypted so that
they do not appear in plain text in the query results)

2. The secret to encrypt/decrypt the strings is stored in a key file inside the
data directory.

3. The data directory path is obtained by using config services.

4. The config services depend on the liquibase/database migration to finish.

5. The liquibase migration script introduced in v3.1 to secure the passwords
stored in DB depends on the config service to obtain the key file for
encrypting the passwords. And the config service depends on liquibase.
This completes a full circle causing the deployment to fail.

6. Resolved the tie by querying the database for the data directory setting,
if any. This way liquibase migration script does not have to depend on the
config service.